Implements loading of 3D model in obj file

+++ PlaneSimulator/Toolkit/Math/MathUtil.cs
@@ -4,13 +4,13 @@
     {
         public const float ZeroTolerance = 1e-6f;
 
-        public const float Pi = 3.141592653589793239f;
+        public const double Pi = 3.141592653589793239f;
 
-        public const float TwoPi = 6.283185307179586477f;
+        public const double TwoPi = 6.283185307179586477f;
 
-        public const float PiOverTwo = 1.570796326794896619f;
+        public const double PiOverTwo = 1.570796326794896619f;
 
-        public const float PiOverFour = 0.785398163397448310f;
+        public const double PiOverFour = 0.785398163397448310f;
 
         public static bool NearEqual(double p1, double p2)
         {
